Clean up join building nits in cumulative metrics conversion
The cumulative metrics logic for converting from dataflow to sql
plans has a number of minor confusing points, which are making it
harder to reason about which pieces of this logic need to be moved
when we fork the join building into the SqlQueryPlanJoinBuilder
class. This commit tidies those things up a bit.