Define abstract `DbtTransformRule` for mapping dbt MetricNode values to Model elements
We want to be able to build a UserConfiguredModel from a series of rules.
Each rule will map a value from a MetricNode, or the MetricNode's sub-
elements, to it's UserConfiguredModel counterpart. However, the
UserConfiguredModel elements have a number of "required" values. Thus
we cannot build the UserConfiguredModel element directly in any given
rule because each rule will deal with one mapping. So instead, slowly
build and modify dicts representing UserConfiguredModel elements via
the rules, and then later we'll build the elments from the dicts.