Add for basic dbt transformation rules for building data sources
DbtTransformRules are intended to map one thing in dbt MetricNodes
(or their sub elements) to one thing in the MetricFlow model. This
commit adds four (4) rules for building MetricFlow DataSources from
dbt MetricNodes. Specifically the rules are:
- Map name of a dbt metric's model to a data source name
- Map descritpion of a dbt metric's model to a data source description
- Map database + schema + name of a dbt metric's model to a data source sql_table
- Map database + schema + name of a dbt metric's model to a data source dbt_model

Additionally this commit adds a helper function to the `DbtTransformedObjects`
data class to help ensure the existance of a specific data source mapping.
We can likely do some defaultdict magic later on to get rid of this and
simplify some code.